+++ modules/canvaskit/interface.js
@@ -848,6 +848,32 @@ CanvasKit.onRuntimeInitialized = function() {
     return ta.slice();
   };
 
+  CanvasKit.ImageFilter.MakeDropShadow = function(dx, dy, sx, sy, color, input) {
+    var cPtr = copyColorToWasm(color, _scratchColorPtr);
+    return CanvasKit.ImageFilter._MakeDropShadow(dx, dy, sx, sy, cPtr, input);
+  };
+
+  CanvasKit.ImageFilter.MakeDropShadowOnly = function(dx, dy, sx, sy, color, input) {
+    var cPtr = copyColorToWasm(color, _scratchColorPtr);
+    return CanvasKit.ImageFilter._MakeDropShadowOnly(dx, dy, sx, sy, cPtr, input);
+  };
+
+  CanvasKit.ImageFilter.MakeImage = function(img, sampling, srcRect, dstRect) {
+    var srcPtr = copyRectToWasm(srcRect, _scratchFourFloatsAPtr);
+    var dstPtr = copyRectToWasm(dstRect, _scratchFourFloatsBPtr);
+
+    if ('B' in sampling && 'C' in sampling) {
+        return CanvasKit.ImageFilter._MakeImageCubic(img, sampling.B, sampling.C, srcPtr, dstPtr);
+    } else {
+        const filter = sampling['filter'];  // 'filter' is a required field
+        let mipmap = CanvasKit.MipmapMode.None;
+        if ('mipmap' in sampling) {         // 'mipmap' is optional
+            mipmap = sampling['mipmap'];
+        }
+        return CanvasKit.ImageFilter._MakeImageOptions(img, filter, mipmap, srcPtr, dstPtr);
+    }
+  };
+
   CanvasKit.ImageFilter.MakeMatrixTransform = function(matrix, sampling, input) {
     var matrPtr = copy3x3MatrixToWasm(matrix);