+++ modules/canvaskit/npm_build/types/index.d.ts
@@ -486,6 +486,7 @@ export interface CanvasKit {
     readonly BlendMode: BlendModeEnumValues;
     readonly BlurStyle: BlurStyleEnumValues;
     readonly ClipOp: ClipOpEnumValues;
+    readonly ColorChannel: ColorChannelEnumValues;
     readonly ColorType: ColorTypeEnumValues;
     readonly FillType: FillTypeEnumValues;
     readonly FilterMode: FilterModeEnumValues;
@@ -1943,7 +1944,7 @@ export interface Image extends EmbindObject<Image> {
      *          not supported in JS, so that colorType corresponds to raw bytes Uint8Array.
      */
     readPixels(srcX: number, srcY: number, imageInfo: ImageInfo, dest?: MallocObj,
-               bytesPerRow?: number): Uint8Array | Float32Array | null;
+               bytesPerRow?: number): Float32Array | Uint8Array | null;
 
     /**
      * Return the width in pixels of the image.
@@ -3340,9 +3341,22 @@ export interface FontMgrFactory {
 }
 
 /**
- * See effects/ImageFilters.h for more.
+ * See //include/effects/SkImageFilters.h for more.
  */
 export interface ImageFilterFactory {
+    /**
+     * Create a filter that takes a BlendMode and uses it to composite the two filters together.
+     *
+     *  At least one of background and foreground should be non-null in nearly all circumstances.
+     *
+     *  @param blend       The blend mode that defines the compositing operation
+     *  @param background The Dst pixels used in blending; if null, use the dynamic source image
+     *                    (e.g. a saved layer).
+     *  @param foreground The Src pixels used in blending; if null, use the dynamic source image.
+     */
+    MakeBlend(blend: BlendMode, background: ImageFilter | null,
+              foreground: ImageFilter | null): ImageFilter;
+
     /**
      * Create a filter that blurs its input by the separate X and Y sigmas. The provided tile mode
      * is used when the blur kernel goes outside the input image.
@@ -3372,6 +3386,87 @@ export interface ImageFilterFactory {
     MakeCompose(outer: ImageFilter | null, inner: ImageFilter | null): ImageFilter;
 
     /**
+     *  Create a filter that dilates each input pixel's channel values to the max value within the
+     *  given radii along the x and y axes.
+     *  @param radiusX  The distance to dilate along the x axis to either side of each pixel.
+     *  @param radiusY  The distance to dilate along the y axis to either side of each pixel.
+     *  @param input     if null, it will use the dynamic source image (e.g. a saved layer).
+     */
+    MakeDilate(radiusX: number, radiusY: number, input: ImageFilter | null): ImageFilter;
+
+    /**
+     *  Create a filter that moves each pixel in its color input based on an (x,y) vector encoded
+     *  in its displacement input filter. Two color components of the displacement image are
+     *  mapped into a vector as scale * (color[xChannel], color[yChannel]), where the channel
+     *  selectors are one of R, G, B, or A.
+     *  The mapping takes the 0-255 RGBA values of the image and scales them to be [-0.5 to 0.5],
+     *  in a similar fashion to https://developer.mozilla.org/en-US/docs/Web/SVG/Element/feDisplacementMap
+     *
+     *  At least one of displacement and color should be non-null in nearly all circumstances.
+     *
+     *  @param xChannel RGBA channel that encodes the x displacement per pixel.
+     *  @param yChannel RGBA channel that encodes the y displacement per pixel.
+     *  @param scale    Scale applied to displacement extracted from image.
+     *  @param displacement The filter defining the displacement image, or null to use source.
+     *  @param color   The filter providing the color pixels to be displaced, or null to use source.
+     */
+    MakeDisplacementMap(xChannel: ColorChannel, yChannel: ColorChannel, scale: number,
+                        displacement: ImageFilter | null, color: ImageFilter | null): ImageFilter;
+    /**
+     *  Create a filter that draws a drop shadow under the input content. This filter produces an
+     *  image that includes the inputs' content.
+     *  @param dx       The X offset of the shadow.
+     *  @param dy       The Y offset of the shadow.
+     *  @param sigmaX   The blur radius for the shadow, along the X axis.
+     *  @param sigmaY   The blur radius for the shadow, along the Y axis.
+     *  @param color    The color of the drop shadow.
+     *  @param input    The input filter; if null, it will use the dynamic source image.
+     */
+    MakeDropShadow(dx: number, dy: number, sigmaX: number, sigmaY: number, color: Color,
+                   input: ImageFilter | null): ImageFilter;
+
+    /**
+     *  Just like MakeDropShadow, except the input content is not in the resulting image.
+     *  @param dx       The X offset of the shadow.
+     *  @param dy       The Y offset of the shadow.
+     *  @param sigmaX   The blur radius for the shadow, along the X axis.
+     *  @param sigmaY   The blur radius for the shadow, along the Y axis.
+     *  @param color    The color of the drop shadow.
+     *  @param input    The input filter; if null, it will use the dynamic source image.
+     */
+    MakeDropShadowOnly(dx: number, dy: number, sigmaX: number, sigmaY: number, color: Color,
+                       input: ImageFilter | null): ImageFilter;
+
+    /**
+     *  Create a filter that erodes each input pixel's channel values to the minimum channel value
+     *  within the given radii along the x and y axes.
+     *  @param radiusX  The distance to erode along the x axis to either side of each pixel.
+     *  @param radiusY  The distance to erode along the y axis to either side of each pixel.
+     *  @param input     if null, it will use the dynamic source image (e.g. a saved layer).
+     */
+    MakeErode(radiusX: number, radiusY: number, input: ImageFilter | null): ImageFilter;
+
+    /**
+     *  Create a filter using the given image as a source. Returns null if 'image' is null.
+     *
+     *  @param img      The image that is output by the filter, subset by 'srcRect'.
+     *  @param sampling The sampling to use when drawing the image.
+     */
+    MakeImage(img: Image, sampling: FilterOptions | CubicResampler): ImageFilter | null;
+
+    /**
+     *  Create a filter that draws the 'srcRect' portion of image into 'dstRect' using the given
+     *  filter quality. Similar to Canvas.drawImageRect. Returns null if 'image' is null.
+     *
+     *  @param img      The image that is output by the filter, subset by 'srcRect'.
+     *  @param sampling The sampling to use when drawing the image.
+     *  @param srcRect  The source pixels sampled into 'dstRect'.
+     *  @param dstRect  The local rectangle to draw the image into.
+     */
+    MakeImage(img: Image, sampling: FilterOptions | CubicResampler,
+              srcRect: InputRect, dstRect: InputRect): ImageFilter | null;
+
+    /**
      * Create a filter that transforms the input image by 'matrix'. This matrix transforms the
      * local space, which means it effectively happens prior to any transformation coming from the
      * Canvas initiating the filtering.
@@ -3381,6 +3476,14 @@ export interface ImageFilterFactory {
      */
     MakeMatrixTransform(matr: InputMatrix, sampling: FilterOptions | CubicResampler,
                         input: ImageFilter | null): ImageFilter;
+
+    /**
+     *  Create a filter that offsets the input filter by the given vector.
+     *  @param dx       The x offset in local space that the image is shifted.
+     *  @param dy       The y offset in local space that the image is shifted.
+     *  @param input    The input that will be moved, if null, will use the dynamic source image.
+     */
+    MakeOffset(dx: number, dy: number, input: ImageFilter | null): ImageFilter;
 }
 
 /**
@@ -3977,6 +4080,7 @@ export type AlphaType = EmbindEnumEntity;
 export type BlendMode = EmbindEnumEntity;
 export type BlurStyle = EmbindEnumEntity;
 export type ClipOp = EmbindEnumEntity;
+export type ColorChannel = EmbindEnumEntity;
 export type ColorSpace = EmbindObject<ColorSpace>;
 export type ColorType = EmbindEnumEntity;
 export type EncodedImageFormat = EmbindEnumEntity;
@@ -4077,6 +4181,13 @@ export interface ColorSpaceEnumValues { // not a typical enum, but effectively l
      * @param b
      */
     Equals(a: ColorSpace, b: ColorSpace): boolean;
+}
+
+export interface ColorChannelEnumValues extends EmbindEnum {
+    Red: ColorChannel;
+    Green: ColorChannel;
+    Blue: ColorChannel;
+    Alpha: ColorChannel;
 }
 
 export interface ColorTypeEnumValues extends EmbindEnum {